Reorganise Xhibit classes into dedicated package
Moved Xhibit-related code (config, enums, services,
etc.) into a dedicated xhibit package with its own
substructure for clarity. Renamed several classes to
better reflect their responsibilities.

No functional changes.
