--- maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/enums/RecordSheetType.java
@@ -1,6 +1,0 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.enums;
-
-public enum RecordSheetType {
-    TRIAL,
-    APPEAL
-}
