+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/xhibit/enums/RecordSheetStatus.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.enums;
+package uk.gov.justice.laa.maat.scheduled.tasks.xhibit.enums;
 
 public enum RecordSheetStatus {
     PROCESSED,
