+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/xhibit/enums/RecordSheetType.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.laa.maat.scheduled.tasks.xhibit.enums;
+
+public enum RecordSheetType {
+    TRIAL,
+    APPEAL
+}
