+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/xhibit/repository/XhibitAppealDataRepository.java
@@ -1,7 +1,7 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.repository;
+package uk.gov.justice.laa.maat.scheduled.tasks.xhibit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uk.gov.justice.laa.maat.scheduled.tasks.entity.XhibitAppealDataEntity;
+import uk.gov.justice.laa.maat.scheduled.tasks.xhibit.entity.XhibitAppealDataEntity;
 
 public interface XhibitAppealDataRepository extends JpaRepository<XhibitAppealDataEntity, Integer> {
 }
