+++ maat-scheduled-tasks/src/main/java/uk/gov/justice/laa/maat/scheduled/tasks/xhibit/repository/XhibitTrialDataRepository.java
@@ -1,7 +1,7 @@
-package uk.gov.justice.laa.maat.scheduled.tasks.repository;
+package uk.gov.justice.laa.maat.scheduled.tasks.xhibit.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uk.gov.justice.laa.maat.scheduled.tasks.entity.XhibitTrialDataEntity;
+import uk.gov.justice.laa.maat.scheduled.tasks.xhibit.entity.XhibitTrialDataEntity;
 
 public interface XhibitTrialDataRepository extends JpaRepository<XhibitTrialDataEntity, Integer> {
 }
